Move all custom components into components/ directory'
